Refactor SwarmAgent to be with rest of Swarm workflow
